+++ cmd/manager/main.go
@@ -30,6 +30,7 @@ import (
 	"github.com/SlinkyProject/slurm-operator/internal/controller/nodeset"
 	"github.com/SlinkyProject/slurm-operator/internal/controller/restapi"
 	"github.com/SlinkyProject/slurm-operator/internal/controller/slurmclient"
+	"github.com/SlinkyProject/slurm-operator/internal/controller/token"
 	//+kubebuilder:scaffold:imports
 )
 
@@ -161,6 +162,12 @@ func main() {
 		setupLog.Error(err, "unable to create controller", "controller", "SlurmClient")
 		os.Exit(1)
 	}
+	if err = token.NewReconciler(mgr.GetClient()).
+		SetupWithManager(mgr); err != nil {
+		setupLog.Error(err, "unable to create controller", "controller", "Token")
+		os.Exit(1)
+	}
+
 	if err := mgr.AddHealthzCheck("healthz", healthz.Ping); err != nil {
 		setupLog.Error(err, "unable to set up health check")
 		os.Exit(1)