+++ cmd/webhook/main.go
@@ -138,6 +138,10 @@ func main() {
 		setupLog.Error(err, "unable to create webhook", "webhook", "LoginSet")
 		os.Exit(1)
 	}
+	if err = (&webhookv1alpha1.TokenWebhook{}).SetupWebhookWithManager(mgr); err != nil {
+		setupLog.Error(err, "unable to create webhook", "webhook", "Token")
+		os.Exit(1)
+	}
 	//+kubebuilder:scaffold:builder
 	setupLog.Info("starting manager")
 	if err := mgr.Start(ctrl.SetupSignalHandler()); err != nil {