+++ internal/builder/token_secret_test.go
@@ -1,0 +1,95 @@
+// SPDX-FileCopyrightText: Copyright (C) SchedMD LLC.
+// SPDX-License-Identifier: Apache-2.0
+
+package builder
+
+import (
+	_ "embed"
+	"testing"
+
+	slinkyv1alpha1 "github.com/SlinkyProject/slurm-operator/api/v1alpha1"
+	"github.com/SlinkyProject/slurm-operator/internal/utils/testutils"
+	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/utils/ptr"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+	"sigs.k8s.io/controller-runtime/pkg/client/fake"
+)
+
+func TestBuilder_BuildTokenSecret(t *testing.T) {
+	jwtHs256Secret := &corev1.Secret{
+		ObjectMeta: metav1.ObjectMeta{
+			Name: "slurm-jwths256key",
+		},
+		Data: map[string][]byte{
+			"jwt_hs256.key": []byte("foo"),
+		},
+	}
+	type fields struct {
+		client client.Client
+	}
+	type args struct {
+		token *slinkyv1alpha1.Token
+	}
+	tests := []struct {
+		name    string
+		fields  fields
+		args    args
+		wantErr bool
+	}{
+		{
+			name: "default",
+			fields: fields{
+				client: fake.NewClientBuilder().
+					WithObjects(jwtHs256Secret).
+					Build(),
+			},
+			args: args{
+				token: &slinkyv1alpha1.Token{
+					ObjectMeta: metav1.ObjectMeta{
+						Name: "slurm",
+					},
+					Spec: slinkyv1alpha1.TokenSpec{
+						Username:       "foo",
+						JwtHs256KeyRef: testutils.NewJwtHs256KeyRef("slurm").SecretKeySelector,
+					},
+				},
+			},
+		},
+		{
+			name: "not found",
+			fields: fields{
+				client: fake.NewFakeClient(),
+			},
+			args: args{
+				token: &slinkyv1alpha1.Token{
+					ObjectMeta: metav1.ObjectMeta{
+						Name: "slurm",
+					},
+					Spec: slinkyv1alpha1.TokenSpec{
+						JwtHs256KeyRef: testutils.NewJwtHs256KeyRef("slurm").SecretKeySelector,
+					},
+				},
+			},
+			wantErr: true,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			b := New(tt.fields.client)
+			got, err := b.BuildTokenSecret(tt.args.token)
+			if (err != nil) != tt.wantErr {
+				t.Errorf("Builder.BuildTokenSecret() error = %v, wantErr %v", err, tt.wantErr)
+				return
+			}
+			switch {
+			case err != nil:
+				return
+
+			case ptr.Deref(got.Immutable, false) != !tt.args.token.Spec.Refresh:
+				t.Errorf("Immutable = %v , want = %v",
+					ptr.Deref(got.Immutable, false), !tt.args.token.Spec.Refresh)
+			}
+		})
+	}
+}